+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Resources/PackageIconMonikers.imagemanifest
@@ -33,7 +33,5 @@
       <Source Uri="$(Resources)/Xamls/MigratorHelpIcon.xaml" />
     </Image>
   </Images>
-  
   <ImageLists />
-  
 </ImageManifest>
